Completed part 5a exercise 5.2: bloglist frontend, step2
