+++ part5/bloglist-frontend/src/App.js
@@ -1,11 +1,14 @@
 import { useState, useEffect } from 'react'
 import Blog from './components/Blog'
 import blogService from './services/blogs'
+import loginService from './services/login'
 import LoginForm from './components/LoginForm'
 import Notification from './components/Notification'
 
 const App = () => {
     const [blogs, setBlogs] = useState([])
+    const [username, setUsername] = useState('')
+    const [password, setPassword] = useState('')
     const [user, setUser] = useState(null)
     const [errorMessage, setErrorMessage] = useState(null)
 
@@ -15,13 +18,48 @@ const App = () => {
         )
     }, [])
 
+    useEffect(() => {
+        const loggedUserJSON = window.localStorage.getItem('loggedBlogappUser')
+        if (loggedUserJSON) {
+            const user = JSON.parse(loggedUserJSON)
+            setUser(user)
+            // blogService.setToken(user.token)
+        }
+    }, [])
+
+    const handleLogin = async (event) => {
+        event.preventDefault()
+
+        try {
+            const user = await loginService.login({
+                username, password
+            })
+            window.localStorage.setItem(
+                'loggedBlogappUser', JSON.stringify(user)
+            )
+            setUser(user)
+            setUsername('')
+            setPassword('')
+        } catch (exception) {
+            setErrorMessage('Wrong Credentials')
+            setTimeout(() => {
+                setErrorMessage(null)
+            }, 5000)
+        }
+    }
+
+    const handleLogOut = () => {
+        setUser(null)
+        window.localStorage.removeItem('loggedBlogappUser')
+    }
+
     return (
         <div>
             <Notification message={errorMessage} />
-            {user === null && <LoginForm user={user} setUser={setUser} setErrorMessage={setErrorMessage} />}
+            {user === null && <LoginForm handleLogin={handleLogin} username={username} setUsername={setUsername} password={password} setPassword={setPassword} />}
             {user !== null && <div>
                 <h1>blogs</h1>
-                <p>{user.name} is logged in </p>
+                <p>{user.name} is logged in <button onClick={handleLogOut}>logout</button></p>
                 {blogs.map(blog =>
                 <Blog key={blog.id} blog={blog} />)}
             </div> }
