+++ part5/bloglist-frontend/src/components/LoginForm.js
@@ -1,29 +1,4 @@
-import { useState, useEffect } from 'react'
-import loginService from '../services/login'
-
-const LoginForm = ({ user, setUser, setErrorMessage }) => {
-
-    const [username, setUsername] = useState('')
-    const [password, setPassword] = useState('')
-
-    const handleLogin = async (event) => {
-        event.preventDefault()
-
-        try {
-            const user = await loginService.login({
-                username, password
-            })
-            setUser(user)
-            setUsername('')
-            setPassword('')
-        } catch (exception) {
-            setErrorMessage('Wrong Credentials')
-            setTimeout(() => {
-                setErrorMessage(null)
-            }, 5000)
-        }
-    }
-
+const LoginForm = ({ handleLogin, username, setUsername, password, setPassword }) => {
     return (
         <form onSubmit={handleLogin}>
             <h1>Log in to application</h1>
